+++ test/Text/CommonMark/SubSpec.hs
@@ -5,14 +5,16 @@ module Text.CommonMark.SubSpec (spec) where
 import Data.Text
 import Test.Hspec
 
+import CMark (Node)
 import Text.CommonMark.QQ
 import Text.CommonMark.Sub
 
-spec :: Spec
-spec = do
-    specify "extractSection" $
-        extractSection 2 "Section to extract" [doc|Doc title
-=====================================
+ciEq :: Text -> Text -> Bool
+ciEq t t' = toLower t == toLower t'
+
+source :: Node
+source = [doc|Doc title
+=======================
 
 Lorem ipsum dolor sit amet, ne omnes vitae mel, alterum invenire duo ut,
 ad deterruisset comprehensam qui. At ius ludus iracundia vituperatoribus.
@@ -39,7 +41,10 @@ Other section
 Pro fuisset deserunt consulatu at, no epicuri quaestio vix,
 case scripserit at vis. No eam alia facilisis. At quo oratio omnium,
 nibh malis dolores an eam.
-|] `shouldBe` [doc|
+|]
+
+expected :: Node
+expected = [doc|
 Section to *extract*
 --------------------
 
@@ -57,6 +62,19 @@ Vivendo assentior mnesarchum ius ei, novum aperiam habemus est in,
 ea agam putant consectetuer sed.
 
 |]
+
+spec :: Spec
+spec = do
+    specify "extractSection" $ do
+        extractSection 2 (==) "Section to extract" source `shouldBe` expected
+        extractSection 3 (==) "Section to extract" source `shouldNotBe` expected
+        extractSection 2 (==) "section to extract" source `shouldNotBe` expected
+        extractSection 3 (==) "section to extract" source `shouldNotBe` expected
+        -- Case insensitivity
+        extractSection 2 ciEq "Section to extract" source `shouldBe` expected
+        extractSection 3 ciEq "Section to extract" source `shouldNotBe` expected
+        extractSection 2 ciEq "section to extract" source `shouldBe` expected
+        extractSection 3 ciEq "section to extract" source `shouldNotBe` expected
     specify "matchesHeading" $ do
         [node|# Title|] `shouldSatisfy` matchesHeading 1 (==) "Title"
         [node|### Title|] `shouldSatisfy` matchesHeading 3 (==) "Title"
@@ -67,7 +85,7 @@ ea agam putant consectetuer sed.
         [node|# Title|] `shouldNotSatisfy` matchesHeading 2 (==) "Title"
         [node|# Title|] `shouldNotSatisfy` matchesHeading 1 (==) "Wrong title"
         [node|# Title|] `shouldNotSatisfy` matchesHeading 1 (==) "title"
-        let ciEq t t' = toLower t == toLower t'
+        -- Case insensitivity
         [node|# Title|] `shouldNotSatisfy` matchesHeading 1 (==) "title"
         [node|# Title|] `shouldSatisfy` matchesHeading 1 ciEq "title"
         [node|# Title|] `shouldNotSatisfy` matchesHeading 2 (==) "title"