+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Version 0.2.0
 
 To be released.
 
+ -  Added `-i`/`--ignore-case` option for case insensitive match.
+
  -  The signature of `Text.CommonMark.Sub.extractSection` function was changed
     from `Level -> Text -> Node -> Node` to
     `Level -> (Text -> Text -> Bool) -> Text -> Node -> Node`; the second